Rename cdatamatch to pdatamatch
